Limit countMemory to 4GB on cygwin32
If cygwin32 is running on a 64-bit windows with more than 4GB of ram
/proc/meminfo will report the true amount of ram the system has, but only 4GB
will be accessible and that value is what our runtime (correctly) reports.

Our cygwin32 box was upgraded from 4GB to 8GB of ram which caused countMemory
to start getting the wrong value. JIRA issue 83 has more details from the last
time we ran into this: https://chapel.atlassian.net/browse/CHAPEL-83. Last time
we just had IT re-limit the machine to 4GB of ram, but this takes a cleaner
approach and limits the amount of ram countMemory reports to
`min($memFreeInBytes, 0xffffffff)` for cygwin32.