Add a future about domain assignment that changes stridedness
This test currently crashes at runtime due to an infinite recursion.  Brad
has a change in progress that should fix that crash, but it gets a runtime
halt.  I think we should catch this case at compile time and report an error
then instead, similar to what we do for a rank mismatch or idxType mismatch.