Remove the "old" perfkey from the MiniMD graph
We no longer run this configuration and it's not of much interest to us
anymore.

Removing this key covers up a gengraphs bug that's currently causing the "old"
key to replace an actual key for chapcs perf graphs.

For chapcs testing, the old MiniMD config was never run, so we didn't have a
.dat file for it. However, gengraphs doesn't create dummy data for the key so
it was "stealing" the actual data from another key.

Unfortunately this is a hard bug to fix with the way gengraphs currently
collates .dat files, so for now I'm just working around it and will work on
fixing gengraphs in the future. Really, I'd like to rewrite most of gengraphs
and add some actual error checking, but I probably won't get time to do that :(