Update valgrind-localhost nodefile
Valgrind testing has been moved from chap11 to chapcs* so switch from
specifying the host as "chap11" to "localhost" (paratest didn't use to support
"localhost", but it does now.) Also bumps the number of instances to 12, which
is half the number of physical cores on a chapcs machine.

This also adds chap11 back to the all-cf file since it won't be used to do
valgrind testing anymore.