Add rank mismatch conditional with runtime halt to avoid dyn dispatch problems
The previous version of this file had problems because our dynamic
dispatch resolution would apparently resolve it sometimes with domains
of rank 1 and 2 for some reason.  I haven't been able to track down
why and meanwhile this noise is getting in developers' ways, so I'm
adopting this workaround for the time being.