Revert fix for Kushal's inter-dmap assignment case
This fix from PR #3423 caused our compiler to generate a (never
called) infinitely recursive routine which results in a clang warning
(treated as error for developers and in the test system) for clang
7.3.0 or later.  I've been trying to put a quick fix together, but
have been thwarted in each attempt so far by dynamic dispatch
resolution and entanglements in the domain map framework.  It still
seems as though a fairly simple solution should not be difficult, but
with the limited cycles I've had to put into it, I haven't been able
to crack it, and it seems like reverting this is the better thing to
do for now given the impacts of the warnings on clang developers.

For future reference, fixes to this future should be tested with clang
7.3.0 on test/release/examples/hello4-datapar-dist.chpl before
committing to make sure that they haven't run afoul of this issue
again.