Start prototyping standard library of int-uint operators
Following up on some ideas I've had for some time and which were
described in my MS6 slides, this adds an example showing how
operators on int/uint pairs could be provided through standard
libraries as a means of overriding the default error behavior.
It relies on marking the original routines as being provided
by the compiler (using a pragma that seems like it should really
be renamed "last resort" or "error fallback case" or something
else implying that it is treated as such in resolution).