Refactor extern C block typedef logic
Typedefs need to be handled both as declarations and
in convertToChplType. So added a new function, convertTypedef,
that does the old logic but can now do it when processing
e.g. a function argument that is a typedef.

(cherry picked from commit 5740b18804dac0c89c28622b5b5364b0c4cd32a9)