Extend table of primes to larger sizes, trim a couple values
The table here was generated with the following program:

use GMP;
use Types;

config const factor = 2.0;
var n:uint = 0x60;

write("(23, 53");
while true {
  // Get the next best prime number.
  const allocSize = chpl_here_good_alloc_size(n:int);
  var prime = prev_prime(allocSize);
  write(", ");
  write(prime);
  n = (factor*n):uint;
  // Break on a number > 2**60
  if n >= 2**60 then break;
}
writeln(");");

proc prev_prime(n:int)
{
  var num1 = new BigInt(0);
  var num2 = new BigInt(0);
  var sub:int = 1;
  while true {
    var target = n - sub;
    if target < 0 then break;
    num1.set_si(target);
    num2.set_si(target);
    num2.nextprime(num1);
    var got = num2.get_si();
    if got < n then return got;
    sub += 1;
  }
  return 0;
}

I experimented with the build-associative test and
 * varying the factor
 * switching to numbers close to 2**i
 * taking into account 16-bytes per element

but I did not find a significant improvement on the values that
were in master.

This change makes two changes to the table of primes:

1) Extends it to values close to 2**60, since before, associative
   arrays of larger than something like 2 G elements would fail.
2) Slightly decrease some of the values to a smaller prime,
   in order to fit in nicely with the space that jemalloc will round up
   allocations into.

In my experiments with the build-associative test, this change does not
significantly affect performance, although it seems to provide a modest
improvement in some cases.