Move param checks for domain assignment out of dsiSetIndices and into '='
Move checks to ensure rank and idxType match for domain assignments into
the domain '=' function.  Add another check that ensures stridable domains
are not assigned to unstridable domains. This was previously allowed if the
stridable domain had stride==1.

The plan is to add a cast from stridable domains to unstridable.  A regular
(unsafe) cast could simply drop the stride without checking it, while a
safeCast can verify that the stride is 1 first.