Make assigning a stridable range to non-stridable range a compiler error
Assigning a stridable range to a non-stridable range is now a compiler error.

Added a cast that can make a stridable range into non-stridable by discarding
the stride even if it is not 1.

Added a safeCast that can make a stridable range into a non-stridable range,
but will halt if stride was not 1.

Fixed a mistake in the domain _cast and safeCast functions that became
apparent with the new range cast functions.  They were causing stridable
ranges to be assigned to non-stridable by mistake.

Updated a .good file to expect the compiler error instead of a halt for
assigning a stridable range to non-stridable with non-unit stride.

Added safeCasts to StencilDist, BlockDist, and BlockCyclicDist in a few places
where they were assigning a stridable range with unit-stride to a
non-stridable range.