Repair sparse subdomain runtime types, dmapped
The previous version changed the arguments to
chpl__buildSparseDomainRuntimeType to only have a domain argument (instead of a
distribution and a domain). I think that was a mistake, since the arguments for
this function define the runtime type for a sparse domain. The runtime type for
a domain normally includes a distribution.

I don't know for certain why this change was made originally, but my guess is
that the purpose was to get the distribution used by a sparse subdomain to be
determined by the original domain/distribution instead of always being a
defaultDist. I believe I have found a different way to do that (I added
_domain.defaultSparseDist which just returns _value.dist).

 * added _domain.defaultSparseDist
 * adjusted the parser to call _domain.defaultSparseDist instead
   of changing arguments to chpl__buildSparseDomainRuntimeType
 * restored original arguments to chpl__buildSparseDomainRuntimeType
 * removed dsiNewSpsSubDom since e.g. Block.dsiNewSparseDom
   seems to cover it adequately.
 * Adjusted Block.dsiNewSparseDom to actually create a
   new SparseBlockDom instead of a DefaultRectangularDom.

Testing

 * test/sparse/parallel/sparse-csr-simple-dom-manips.chpl now works
 * blockSparse1.chpl still works
 * (but note blockSparse2.chpl is not currently working,
    although it didn't work before these changes either).