Minor vararg-related fixes in compiler
* Replace explicit calls to 'delete' etc. on AST nodes
(inserted by a collaborator)
with the standard node->replace().

* Added a comment on FnSymbol::variableExpr field.
Taken from compilerOverview.pdf.

* All code in handleSymExprInExpandVarArgs() is gated by two conditionals.
I verified and documented, using INT_FATAL, that the inner conditional
always fires.
Also documented when the outer condition may be false.
This was illustrated by this one test in our suite:

  test/functions/deitz/varargs/test_varargs3.chpl

* Moved "newFn->partialCopySource = this;"
from copyInnerCore() to partialCopy() - for clarity:
do not do it where it is not needed.

Passes full linux64 test suite.