If we might have sub-locales, use multiple ddata blocks for arrays.
==========
Changes in DefaultRectangular.chpl

Add support for multiple ddata chunks when the locale model's locales
may have sublocales, based on the param defRectSimpleDData, whose value
in turn reflects localeModelHasSublocales, provided by the locale model.
If sublocales are not present, store the data in single ddata chunk as
before, in the data1 (was data) and shiftedData1 (was shiftedData)
fields.  But if sublocales may be present, instead allocate chunks of
ddata space, referencing those through dataVec and shiftedDataVec.  If
the array has fewer than (param) defRectArrMultiDDataSizeThreshold
elements or only one sublocale exists, just allocate one such chunk.
Otherwise choose a dimension on which to chunk (using DSIUtil range
chunking) and split that dimension across multiple chunks.

Besides dataVec and shiftedDataVec there are several metadata fields
used in the chunk computations:
  mdParDim:       array is chunked on this dimension
  mdNumChunks:    number of chunks
  mdRLo:          original dom.dsiDim(mdPardim).low
  mdRHi:          original dom.dsiDim(mdPardim).high
  mdRLen:         original dom.dsiDim(mdPardim).length

A variety of other changes flow from this.  The dataVecs have to be
allocated and freed, obviously.  The old parentheses-less method theData
returned a ref to data or shiftedData depending on whether or not we
were using 0-based indexing.  Instead we now have theDataChunk(i), where
i is a chunk index (always 0 when defRectSimpleDData), which returns the
appropriate chunk, and theData(i) (with parentheses) which returns the
given array element.  I split the array iterators into two distinct
groups for the simple-ddata and multi-ddata cases, for coding clarity.
Those for the simple-ddata case are largely unchanged from what was
there before, while the multi-ddata iterators are all new.

initialize() has quite a bit of new code, to do the chunking and set up
all the meta-data.

Two utility functions mdInd2Chunk() and mdChunk2Ind do the work of
translating back and forth between array indices and chunk indices.
getDataIndex() now returns a tuple (chunk index, index within chunk)
instead of a scalar (index within data).  However, it can also compute
unchunked indexes even in the multi-ddata case, because those are needed
here and there.  I also parameterized it to be able to compute "shifted"
indexes even with earlyShiftData, because those are also needed here and
there and this seemed cleaner than getUnshiftedDataIndex() (which I
removed).

dsiReindex(), dsiSlice(), dsiRankChange(), and dsiReallocate() all have
changes to compute the multi-ddata metadata for the new array.

The RADopt support includes new meta-data fields and methods matching
those in the arrays for multi-ddata.  In addition there is now a real
lock in the locRADCache class, because filling that can no longer be
done redundantly due to the need to allocate multi-ddata dataVecs.
(This is used in the distributions.)

Array I/O and bulk transfer all have extensive changes to support
multi-ddata arrays.  Currently for multi-ddata arrays we can only do
direct bulk transfer on multi-ddata arrays if the source and target are
chunked exactly the same, and we can only do strided bulk transfer if
both arrays are single-ddata.

While here, I also reworked computeFactoredOffs() a little bit to
clarify the code.  Also, I added idxSignedType just as a shorthand.

==========
modules/internal/ChapelArray.chpl

Add doiUseBulkTransfer() and doiUseBulkTransferStride() calls to
chpl__useBulkTransfer() and chpl__useBulkTransferStride(), respectively.
Whether or not we can do bulk transfer, strided or not, now depends
partly on whether and how the source and target arrays are chunked for
multi-ddata.  So, we have to check that.

==========
modules/dists/BlockDist.chpl
modules/dists/CyclicDist.chpl
modules/dists/SparseBlockDist.chpl
modules/dists/StencilDist.chpl

The locRAD stuff all needs a stridable param now, because multi-ddata
indexing uses that to save operations in the non-strided case.

Augment getDataIndex to handle multi-ddata sub-arrays.

Replace the "known, benign race" in initializing LocArr.locRAD.RAD()
elements with an actual lock, because now we may have to allocate the
dataVec and shiftedDataVec fields and that must be done exactly once.

Add a .oneDData method that tells whether all the blocks of a
distributed array are in single-ddata sub-arrays.  Use this in
doiUseBulkTransfer() and doiUseBulkTransferStride() in each
distribution to tell whether or not we should do bulk transfer.
Basically, we can do it (using the existing code) if the arrays have
exactly the same domains or if all their sub-arrays are single-ddata.
Remove some assert(<subArray>.oneDData) assertions that are no longer
needed as a result.  Fix bulk transfer getDataIndex() calls to ask for
single-ddata indexes.

Allow bulk transfer for multi-ddata sub-arrays only when either the
sub-arrays are blocked exactly the same (for direct bulk) or all the
sub-arrays ahev just a single ddata chunk (for direct or strided).

While here, fix a few typos.

==========
modules/internal/ChapelLocale.chpl

Assert that the locale arrays are single-ddata, because we need to
access their raw data in order to do a home-grown bulk copy.

Also, fix a typo.

modules/internal/localeModels/flat/LocaleModel.chpl

I changed a compile-time error to an execution-time halt in the flat
locale model, so that the presence of chpl_getSubloc() calls in the
debug support in DefaultRectangular didn't kill compilation even though
they weren't called.  It may be that this is no longer needed, though;
the code may have evolved to the point that we don't ask for the subloc
unless locModel!=flat.

==========
Notes for the future:

There are still bugs in here, mostly I suspect when sliced or reindexed
or otherwise reshaped arrays are in use.  A "stress" testing run with
defRectArrMultiDDataSizeThreshold=32 producxed 90 or so failures.

It would be nice if we could cover more multi-ddata arrangements in
strided bulk transfer.