Update dsiAccess in various places
Because of the risk of returning an address of a local temporary,
implementations of dsiAccess should only make a const-ref return version
if there is also a value return version.

The where clause
  where !shouldReturnRvalueByConstRef(eltType)
should be on the value version and the where clause
  where shouldReturnRvalueByConstRef(eltType)
should be on the const ref version.

Other distributions only have a ref return version, which is OK.
It would also be OK to only have ref and value return versions,
since there is no risk then of returning an address of a local.
However, when I initially added these, I was concerned about the
extra copies generated by the value return versions, and so I
added const ref return versions in most places.

While investigating this issue with UserMapAssoc, I realized
that the value version was never running, so fixed that.