Provide an updated version of the chpl.lang file for Andre Simon's highlighter
The version we had was compatible with the 2.x release of the highlighter, but
not 3.x when it switched to using Lua for the backend scripts.  I took a look
at the c.lang file in the most up-to-date version of highlight (3.28) and
modified it to utilize the information in the old chpl.lang file.  Also updated
the README to mention the latest version and the new syntax to highlight a file