Allow dygraph annotations to work for days where there is no data
Dygraph annotations require an "x-value" to attach to. i.e. there must be an
actual date-point in order for an annotations to be attached. Previously this
meant that we couldn't apply annotations for dates where the data was missing.

This patch fills in missing dates with "empty" dates (adds a day with blank
data) so there's something for the annotation to attach to.