Add c_memmove as a simple wrapper over memmove
Add c_memmove(dest, src, n) as a simple wrapper of the C function memmove().
The only difference is that n is an integral instead of size_t. This function
takes care of the safeCast to size_t.

Then use this function in String.chpl instead of relying on a private memmove
wrapper (that didn't handle casting)