Add c_memcpy as a simple wrapper over memcpy
Add c_memcpy(dest, src, n) as a simple wrapper of the C function memcpy().
The only difference is that n is an integral instead of size_t. This function
takes care of the safeCast to size_t.

Then use this function in String.chpl instead of relying on a private memcpy
wrapper (that didn't handle casting)