Add c_memcmp as a simple wrapper over memcmp
Add c_memcmp(s1, s2, n) as a simple wrapper of the C function memcmp().  The
only difference is that n is an integral instead of size_t. This function takes
care of the safeCast to size_t.

Then use this function in String.chpl instead of relying on a private memcmp
wrapper (that didn't handle casting)