Simplify and speed up chpl_executeOn().
For chpl_executeOnNB()s that are network-local we've been (ab)using
chpl_comm_execute_on_NB() to initiate an asynchronous local task.  Here
we update the tasking layer interface to do this directly instead.

Specifically, we add chpl_task_taskCall() and chpl_task_taskCallFTable(),
which create tasks to call a pointed-to function or a function out of
chpl_ftable[], respectively.  The former is implemented in each of the
tasking layers, while the latter is just a thin shim in chpl-tasks.h.  As
part of doing this I also adjusted chpl_task_startMovedTask() so that it
and chpl_task_taskCall() share most of their code, since the two are
quite similar.  (In fact, when chpl_comm_execute_on_NB() was being used
to do asynchronous local calls chpl_task_startMovedTask() was the tool it
used to do so.)

With these available, I adjusted chpl_executeOnNB() in the flat and numa
locale models to use chpl_task_taskCallFTable() in the local, parallel
case.  While in that code I also did some cleanup, reordering an if-stmt
in numa so that it tested for local instead of nonlocal (more natural
and now matches flat), fixing white space here and there, and removing
some superfluous or incorrect comments.

I believe there is a performance improvement that could still be made in
this code.  executeOn body argument bundles are dynamically allocated by
the calling code, then freed by that code after the task creation
request is made to the runtime.  Because the bundle free() occurs right
away, in the executeOnNB case probably before the task actually runs,
the runtime has to make a copy of the bundle for the task to use.  If we
could change the compiler/runtime contract so that the emitted code
allocated the bundle and the runtime was responsible for freeing it, we
could save the extra allocate/copy/free in the executeOnNB case.