Fix failure to report ambigous value-version of a ref-pair
The below code was not triggering an ambiguity error in function resolution due
to some problems with how the ref-pair support was initially implemented.

    var global = 0;
    proc access() ref { return global; }
    proc access() { return global; }
    proc access() const ref { return global; }

    var x = access();  // value version should be ambiguous
    writeln(x);

Instead, the ref version was being called (because disambiguateByMatch
was returning NULL for ambiguous but the code assumed it meant
no suitable function was found).

This PR fixes this problem. In order to do so, it separates the logic for
finding ref return and not-ref return candidates from disambiguateByMatch. The
problem with the disambiguateByMatch approach is that disambiguateByMatch
returns NULL for either

 * no matching candidates, or
 * ambiguity

Other function resolution code checks for ambiguity by inspecting the number
of candidates. So, this PR adds a function, splitCandidates, that separates the
candidates into ref and non-ref lists. Additionally, it carefully handles the
case when disambiguateByMatch returned NULL but either of these lists has > 0
elements.