Adding a domain align operator
Nikhil pointed out that domains seemed to be lacking an 'align'
operator, which seems correct and strikes me as simply being an
oversight.  Here I've added it by simply cloning the logic for the
'by' operator on domains and added a simple test the captures the case
that Nikhil sent in by mail.