Only insert implicit 'use ChapelStandard' for top-level user modules
Historically, we've inserted 'use ChapelStandard' into every user
module.  However, this seems inappropriate for nested user modules
because it causes them to favor symbols in internal/standard modules
over those in their parent module which seems unintuitive/ incorrect.
As a simple example, imagine that 'foo' is a symbol defined in an
internal/standard module that ChapelStandard references.  Consider:

    module M {
      // compiler-inserted 'use ChapelStandard;'

      var foo: int = 3;

      module Inner {
        // compiler-inserted 'use ChapelStandard;'

        writeln(foo);
      }
    }

Here, the writeln() of foo in the inner module would refer to the
'foo' defined by ChapelStandard rather than the obvious one in
the outer module.  Meanwhile if we only insert the use into the
outermost module:

    module M {
      // compiler-inserted 'use ChapelStandard;'

      var foo: int = 3;

      module Inner {
        writeln(foo);
      }
    }

The code works as you'd guess and other symbols within ChapelStandard
would also be available to Inner through its parent module's use of
ChapelStandard.