Fix -b -t with FIFO+Gasnet on 1 locale
PR #3729 moved main into a new thread, so it could get
a stack allocated by the runtime. However, the support
for lock reporting was setting thread-local storage for
main, but not in the new thread running chpl_main.

This patch just moves the call to initializeLockReportForThread from
chpl_task_init to do_callMain.