Reorganize multipleDomains2 .good files.
The only .good file for multipleDomains2 with CHPL_LOCALE_MODEL=numa was
multipleDomains2.lm-numa.good, which matched the program output when
CHPL_COMM=none.  But it didn't match when CHPL_LOCALE_MODEL=numa and
CHPL_COMM!=none; we actually didn't have a .good file at all for that
configuration.  For CHPL_LOCALE_MODEL=flat we had multipleDomains2.good
for comm=gasnet and multipleDomains2.comm-none.lm-flat.good for
comm=none.  This was all a bit confusing.

Here, create a new .good for CHPL_COMM!=none, CHPL_LOCALE_MODEL=numa.
Then rearrange the other existing .good files so as to reach this state:
  CHPL_COMM     CHPL_LOCALE_MODEL  .good
  =========     =================  =====
  none          flat               multipleDomains2.comm-none.good
                numa               multipleDomains2.comm-none.lm-numa.good
  gasnet,ugni   flat               multipleDomains2.good
                numa               multipleDomains2.lm-numa.good

I tested this with all 6 of these configurations.