Adjust compiler, locale models for direct on statement invocation
Before this change, the compiler always allocated an argument
bundle for an on statement execution even if that on statement
was to run locally.

After this change, instead of unconditionally emitting code
to allocate the argument bundle and call chpl_executeOn, the
compiler emits a code sequence like this:

  if (chpl_doDirectExecuteOn(targetLocale)) {
     call_on_body(args...);
  } else {
     // the old sequence, as before
     allocate arg bundle
     chpl_executeOn(targetLocale, arg bundle)
     free arg bundle
  }

The chpl_doDirectExecuteOn function is implemented in the
locale models. It needs to be provided by the locale model
because it is up to the locale model what the sublocale values
mean and it is up to the locale model what the policy is about
when a task needs to be moved to another locale/sublocale.