Move on-statement body fences to wrapper function
In the case of a direct on-statement invocation, running the on statement
is just doing a function call. Since it is just a function call, fences
for --cache-remote are not necessary.

This patch moves the fences added to on-statement functions from the
on-statement function itself to the wrapper. That way, in a direct
on-statement invocation, the fences are not run, but if the on-statement
could not be handled directly (e.g. it is on a different node or
sublocale) then the fences are run as part of the wrapper.

The mechanism that this patch uses to do this is simply to add flags to
the on-statement function in createTaskFunctions indicating if a start or
finish fence is needed. Then, in parallel.cpp, the code creating the
wrapper function will add these fences if the appropriate flag is set.