Retire timeouts for thread ring
These timeouts originally date back to when we were using 'fifo'
tasking which took quite a bit longer for this benchmark.  Even after
the switch to 'qthreads' by default, we kept the timeout because
prgenv-cray didn't support qthreads at the time and pgi doesn't
support native atomics, so those two configurations continued to be
slow.  Now that cce does support qthreads and we're not doing pgi
performance testing, it should be OK to remove these timeouts.
Spot-checking the graphs, everything seems pretty solidly at or below
200 seconds across our test machines.