Downsizing minimal modules.
My new take on minimal modules is to provide bare minimals.
In the future we intend to provide separate modules
with additional functionality, e.g. int, string, tuple,
maybe even tasks and foralls, that minimal-modules programs
can "use" explicitly. Such modules could be created incrementally
as developers develop them for their own needs.

Turns out that only the _ref type is really needed for bare minimals.
E.g. because resolveExpr() relies on it when handling a SymExpr.

After trying various ways to mock up strings in a minimal way,
I realized that strings could not be quite minimal - they entail
at lesst the c_ptr type, constructor, autoCopy and autoDestroy.

Then it occurred to me that I can just get rid of it altogether
if strings are not used. In other words, if the user program
and/or minimal modules do not define the string type,
the compiler can handle that. If they do define, the compiler
should be able to handle that, too, although I have not tried.

For now -- "out of the box" minimal modules do not support strings.
So use c-strings instead, as illustrated by the minMod tests.

I also removed tuples, in the spirit of minimality.
So the printf prototypes use fixed number or formals instead of varargs.
Once my #3838 is merged, we may be able to use varargs
in limited situations again.

Discussed with Brad.