Updated pidigits versions to support n != 0, mod 10; ledrug version
Lydia pointed out last week that our pidigits versions only printed
correctly when 10 divided n evenly and that the competition requires
running correctly for an input of 27.  I'm wondering if this changed
at some point in the past, as I thought that the default of 50 was
based on the website, but perhaps I'm misremembering (she mentioned
that some of the versions on the site also don't work in that case,
so maybe we just learned from the wrong ones).

Anyway, this puts a cleanup loop after the main loop to print out
the rest of the line in the event that n isn't a multiple of 10.

I also did some cleanup to the ledrug version to make it more similar
to the release version which we reviewed in last week's performance
meeting.