Adding .perfexecopts files
In order to test both the -n=10000 and -n=27 cases, I had to remove
EXECOPTS because it doesn't support multiple lines (?!) which also is
used for the default PERFEXECOPTS (confusing!  Did I code this up?)
so I had to replace .perfexecopts for these tests to make them work
as before.