Resolve default type fields in aggregate return type
In for example

    proc foo(): range { }

range is a generic aggregate type with a default value for its generic
fields. These were previously resolved by resolveDefaultGenericType in
handling the _statementLevelSymbol call. But, I removed the
_statementLevelSymbol call - so this change is necessary to recover the
original behavior.

Also changes the error message for functions which declare a generic
return type to be more clear and to allow compilation to continue some.
For this reason - the returnGeneric.good files changed and got additional
errors. As always, there is a tradeoff between continuing compilation and
producing possibly confusing follow-on errors and stopping compilation
but only reporting one error at a time.