Remove record chpl_main_argument for --minimal-modules
This change makes minimal modules more minimal.

It was easy to have the compiler accomodate the absence of declaration
of chpl_main_argument - see the change in buildDefaultFunctions.cpp.

As a result, the generated C code defines chpl_gen_main() taking no arguments.
Which contradicts the prototype of chpl_gen_main() in the runtime.
To match that prototype, we create record chpl_main_argument
and add an appropriate argument to chpl_gen_main() right before codegen.