Adjusted my future on negative reference count
Turns out that valgrind DOES report an error on my previous test.

Clarification: valgrind does not report error if CHPL_MEM is jemalloc and
there are no valgrind headers installed on the system. So need to run it
with CHPL_MEM=cstdlib or install the valgrind-devel package.

I adjusted the test accordingly, and simplified it too.

From the adjusted .future:

In this code:

  var bd = new Block(...);
  var dm = new dmap(bd);

* We want dm to wrap bd directly.

* We want bd to remain live afterwards.

Neither is hapenning now. Instead, new dmap(bd):

* clones bd i.e. using .clone() and .dsiClone(), and

* invokes autoDestroy on the original, which decrements
its reference count; Since the ref cnt reaches 0, bd is deleted.