Allow late setting of Block's bounding box (not in modules)
This allows creation of Block objects when the bounding box
is not known yet, then setting it later.

The test file contains both the implementation and a test harness.

The code I listed in the WARNING comment misbehaves under comm=none,
however works fine with --nolocal and gasnet.
Alas the "_distCnt.read() < 0" guard did not fire for me
in this case under comm=none.
The corresponding .future is:

  test/distributions/vass/negative-ref-count-dmap.chpl

We may want to consider productizing this.