GMP.chpl: fix extern proc mpz_conguent_p -> mpz_congruent_p
Fix the name of this extern C function.  This matches the spelling
used to call the function later in the file, in BigInt.congruent_p().

Without this, a program using BigInt.congruent_p() fails to compile.
With it, it works.

(cherry picked from commit a1ea15af095cc7de608d40466449553461f2dc1e)