Stop artificially serializing promoted array assignment
Mike happened to have me look at this routine today and I couldn't see
any reason that we're manually serializing non-rectangular array
assignments in the promotion case, so I removed it and nothing seemed
to break or be the worse for the wear.  While it could be that some
array implementations may still not support parallel iteration, it
should be the array implementation that should warn about that, not a
decision made at this high-level operation.