Stress tests for reductions and reduce intents
reductions-stress-mf.chpl is the test due to Michael that exhibited
the race in reduction that was fixed in #3886. In my experiments
with comm=none, tasks=fifo and the runtime built with CHPL_DEVELOPER=1,
it reports 400..500 errors when I revert #3886.

When the program is compiled with --fast, no errors are reported.
This is perhaps because the read through the shared pointer
is optimized away in that mode.
Yet I am including a version of it, with 10x problem size,
called reductions-stress-fast.

I am also adding a version of ri-3-stress to be compiled without --fast.
While ri-3-stress does not trigger the race of fixed in #3886, I think
it's a good idea to test it without --fast as well.