Change gasnet source directory structure.
(This follows the lead of similar changes to third-party/jemalloc and
third-party/hwloc.)

Rename GASNet-1.26.0. => gasnet-src so that we can avoid adding and
removing new/old source versions when we upgrade.

We used to add the version number to our 3rd party sources.  When we
upgrade versions we would add/commit the new tarball and the untarred
source.  The we would remove the tarball, and switch our makefiles to
use the new version.  Finally, after enough testing, we would remove the
old sources.

While this was nice because it allowed us to switch between two versions
easily and embedded the version name in the directory structure, it
added a lot of bloat to the git history.  We had two commits to add and
remove the tarball, and instead of just committing the diff between
releases, we ended up completing adding the new source in a new dir, and
removing the old source later.  This new setup will allow us to just
commit the diff between releases, which should drastically reduce the
amount stored in the git history.