Add our "traditional" change that allows running a debugger on each child.
This allows us to invoke gdb or lldb in a separate xterm window for each
child of a Chapel program that uses the GASNet amudprun launcher.  The
change is an agglomeration of the following commits applied to the
previous GASNet 1.26.0 package version:

    8780c6cbd8b42411740c03e0ca411fab759cba9d
    07ae945dba21ea646416848b8701c6003b5e9615
    3e27a370fb9ecae6b0f44c69f55cb2d3d6db80c6
    056d38adbff5994767bc5f01e2b102e43497629d
    b8d5eb2e552eeec189192b205d14a79005631277
    72c9cf0daa6b1a4e40abc90f041b8d53f30fbeb8

Conceptually, I applied these changes using the following, where the
'-p4 --directory=...' options were necessary because the amudp_spmd.cpp
file had moved from GASNet-1.26.0/... to gasnet-src/... since those
commits were done:

  git show $(< file-containing-those-commit-shas ) \
  | git apply -p4 --directory=third-party/gasnet/gasnet-src/

In reality it was a bit more complicated because the last commit above
affected more than just the amudp_spmd.cpp file, including some files
that had fewer than 4 components which caused the 'git apply' to bark at
the '-p4'.  As a result I had to manually pluck out and apply just the
part of it that corresponded to that file.  In the future that won't be
an issue because only this commit will need to be re-applied, and with
third-party/gasnet/GASNet-<ver> renamed to third-party/gasnet/gasnet-src
the file path for amudp_spmd.cpp won't change from release to release of
GASNet.