Upgrade jemalloc from 4.0.4 to 4.2.0
I originally tried upgrading to 4.1.0 with #3447, but had to revert to 4.0.4
with #3477 because 4.1.0 didn't have a configure test to allow us to detect if
librt needed to be linked in. This issue was resolved with 4.2.0, so we can
start using it. https://chapel.atlassian.net/browse/CHAPEL-192 has more info on
the issue

4.1.0 was "primarily about optimizations, but it also incorporates a lot of
portability-motivated refactoring and enhancements." 4.2.0 fixes the librt
dependecy issue and has a few other optimizations for internal data structures.

For the changelog see: https://github.com/jemalloc/jemalloc/releases/tag/4.2.0

In my testing I saw a slight improvement for things like miniMD, array-as-vec
operations, nbody, and some other benchmarks that improved with the original
jemalloc switch.

Binary Trees (a benchmark that really measures allocator speed) shows a ~10
performance improvement with 4.2.0